refactor(web): extract onboarding v4 local-default helper
Keep the local/hosted/prod default in a dependency-free module so tests
can cover the matrix without loading analytics.
